Refactor simulator into its own class
